feat: compile server loader
