+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToString/loadable.ts
@@ -33,7 +33,7 @@ export const toHtml: RenderHandler = (jsx, renderer, next) => {
   chunksMap.js = (chunksMap.js || '') + getLoadableScripts(extractor);
 
   for (const v of chunks) {
-    const fileType = extname(v.url!);
+    const fileType = extname(v.url!).slice(1);
 
     if (fileType === 'js') {
       const props = [];
