+++ packages/toolkit/utils/src/index.ts
@@ -37,3 +37,4 @@ export * from './analyzeProject';
 export * from './chainId';
 export * from './version';
 export * from './plugin';
+export * from './routes';
